+++ apk/Changelog.md
@@ -11,6 +11,7 @@ Web Version September 01 2020
 ==============
 
 * Improve Stay on the stream feature, prevent buffer dialog and last video frame from be displayed when the mode starts, prevent screen saver
+* Improve offline chat visual and fix write to chat not showing some times
 
 Web Version August 30 2020
 ==============